requirement.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.8
django-cors-headers==3.7.0
djangorestframework==3.12.4
frida==14.2.18
frida-tools==9.2.5
idna==2.10
mysqlclient==2.0.3
Naked==0.1.31
prompt-toolkit==3.0.18
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
Pygments==2.9.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
watchdog==2.1.2
wcwidth==0.2.5
